requirements.txt
gradio
opencv-python
#torch==2.2.1+cu121
#torchaudio==2.2.1+cu121
#torchvision==0.17.1+cu121
tqdm
transformers
transparent-background
pillow
onnxruntime-gpu
Pillow==9.5.0
pyaml
moviepy